# specs/04_api_spec.yaml
openapi: 3.0.0
info:
  title: MTF Trading API
  version: 0.1.0
paths:
...
  /api/v1/funds:
    get:
      summary: List funds for current user
      responses:
        '200':
          description: List of funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_Fund'
    post:
      summary: Create a new fund
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FundCreate'
      responses:
        '201':
          description: Fund created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_Fund'
  /api/v1/funds/{fund_id}:
    get:
      summary: Get fund details
      parameters:
        - name: fund_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Fund details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_Fund'
    put:
      summary: Update fund details
      parameters:
        - name: fund_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FundUpdate'
      responses:
        '200':
          description: Fund updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_Fund'
    delete:
      summary: Delete a fund
      parameters:
        - name: fund_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Fund deleted
...
components:
  schemas:
...
    ResponseStatus:
      type: string
      enum: [success, error, fail]
    ErrorCode:
      type: string
      enum:
        - AUTH_1001
        - AUTH_1002
        - AUTH_1003
        - AUTH_1004
        - AUTH_1005
        - VAL_2001
        - VAL_2002
        - VAL_2003
        - VAL_2004
        - VAL_2005
        - VAL_2006
        - VAL_2007
        - RES_3001
        - RES_3002
        - RES_3003
        - BIZ_4001
        - BIZ_4002
        - BIZ_4003
        - BIZ_4004
        - RATE_5001
        - RATE_5002
        - SRV_9001
        - SRV_9002
        - SRV_9003
        - SRV_9004
...
    APIResponse:
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          $ref: '#/components/schemas/ResponseStatus'
        data:
          type: object
          nullable: true
        message:
          type: string
          nullable: true
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
        meta:
          $ref: '#/components/schemas/Meta'
        auth:
          $ref: '#/components/schemas/AuthTokens'
        rate_limit:
          $ref: '#/components/schemas/RateLimitInfo'
        timestamp:
          type: string
          format: date-time
...
    Meta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    AuthTokens:
      type: object
      required:
        - access_token
        - expires_in
        - expires_at
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: Bearer
        expires_in:
          type: integer
        expires_at:
          type: string
          format: date-time
    RateLimitInfo:
      type: object
      required:
        - limit
        - remaining
        - reset
        - reset_in_seconds
      properties:
        limit:
          type: integer
        remaining:
          type: integer
        reset:
          type: string
          format: date-time
        reset_in_seconds:
          type: integer
    ErrorDetail:
      type: object
      required:
        - message
      properties:
        field:
          type: string
        message:
          type: string
        code:
          $ref: '#/components/schemas/ErrorCode'

    # --- Generic Wrappers (instantiated for each type) ---
    # Note: OpenAPI 3.0 doesn't support generics directly, so we define explicit types.
...
    APIResponse_Fund:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Fund'
...
    PaginatedResponse_Fund:
      type: object
      required:
        - status
        - data
        - meta
        - timestamp
      properties:
        status:
          $ref: '#/components/schemas/ResponseStatus'
        data:
          type: array
          items:
            $ref: '#/components/schemas/Fund'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/Meta'
        rate_limit:
          $ref: '#/components/schemas/RateLimitInfo'
        timestamp:
          type: string
          format: date-time
...
    Fund:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    FundCreate:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
        description:
          type: string
        strategy_type:
          type: string
          enum: [MTF_SMC_BASIC, LONG_SHORT_EQUITY, MACRO_TACTICAL, MULTI_ASSET]
        asset_classes:
          type: array
          items:
            type: string
        max_risk_per_trade:
          type: number
        default_lot_size:
          type: number
        max_drawdown_threshold:
          type: number
          nullable: true
        max_portfolio_beta:
          type: number
          nullable: true
        gross_exposure_limit:
          type: number
          nullable: true
        net_exposure_limit:
          type: number
          nullable: true
        position_limit_single:
          type: number
          nullable: true
        position_limit_sector:
          type: number
          nullable: true
    FundUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        strategy_type:
          type: string
          enum: [MTF_SMC_BASIC, LONG_SHORT_EQUITY, MACRO_TACTICAL, MULTI_ASSET]
          nullable: true
        asset_classes:
          type: array
          items:
            type: string
          nullable: true
        max_risk_per_trade:
          type: number
          nullable: true
        default_lot_size:
          type: number
          nullable: true
        max_drawdown_threshold:
          type: number
          nullable: true
        max_portfolio_beta:
          type: number
          nullable: true
        gross_exposure_limit:
          type: number
          nullable: true
        net_exposure_limit:
          type: number
          nullable: true
        position_limit_single:
          type: number
          nullable: true
        position_limit_sector:
          type: number
          nullable: true
...